feat: Add user model
The code changes add a new file  that defines the user model. This model includes columns for uid=1000(ezeisraeljohn) gid=1000(ezeisraeljohn) groups=1000(ezeisraeljohn),4(adm),20(dialout),24(cdrom),25(floppy),27(sudo),29(audio),30(dip),44(video),46(plugdev),116(netdev),999(docker), , , , and . The commit message follows the established convention of using a prefix to indicate the type of change, in this case, \feat\ for a new feature.